:construction: refactor header forwarding to reduce conflict surface
